move extHostMain to /common/, have base extension service in /common/ and extend in /node/
